Move SignalNoCallbacksRunning back to shared code, add static EventWaitHandle.Set.
